requirements.txt
requests
python-dotenv
transformers
gradio